User Interface Update
Code was revised for style and commented further/javadocs added.
Additionally, a few bugs were cleared up in the system.
